/tests/curl - Adjust setup of the webserver
In some circumstances `python -m http.server` may take too long to
setup, and we try to access the page with curl before it can be served.

In this change we don't completely eliminate the issue, but by explicitly
binding the webserver to localhost interface we make python webserver to
start almost immediately.